+++ packages/jsii-rosetta/bin/jsii-rosetta.ts
@@ -4,12 +4,13 @@ import * as fs from 'fs-extra';
 import * as path from 'path';
 import * as yargs from 'yargs';
 
-import { TranslateResult, DEFAULT_TABLET_NAME, translateTypeScript, RosettaDiagnostic } from '../lib';
+import { TranslateResult, translateTypeScript, RosettaDiagnostic } from '../lib';
 import { translateMarkdown } from '../lib/commands/convert';
-import { extractAndInfuse, ExtractResult, extractSnippets } from '../lib/commands/extract';
+import { extractAndInfuse, extractSnippets, ExtractOptions } from '../lib/commands/extract';
 import { infuse, DEFAULT_INFUSION_RESULTS_NAME } from '../lib/commands/infuse';
 import { readTablet } from '../lib/commands/read';
 import { transliterateAssembly } from '../lib/commands/transliterate';
+import { trimCache } from '../lib/commands/trim-cache';
 import { TargetLanguage } from '../lib/languages';
 import { PythonVisitor } from '../lib/languages/python';
 import { VisualizeAstVisitor } from '../lib/languages/visualize';
@@ -69,37 +70,31 @@ function main() {
       '(EXPERIMENTAL) mutates one or more assemblies by adding documentation examples to top-level types',
       (command) =>
         command
-          .positional('TABLET', {
-            type: 'string',
-            required: true,
-            describe: 'Language tablet to read',
-          })
           .positional('ASSEMBLY', {
             type: 'string',
             string: true,
             default: new Array<string>(),
             describe: 'Assembly or directory to mutate',
           })
-          .option('log', {
+          .option('log-file', {
             alias: 'l',
-            type: 'boolean',
-            describe: 'Test all algorithms and log results to an html file',
-            default: false,
-          })
-          .option('output', {
-            alias: 'o',
             type: 'string',
             describe: 'Output file to store logging results. Ignored if -log is not true',
             default: DEFAULT_INFUSION_RESULTS_NAME,
           })
-          .demandOption('TABLET'),
+          .option('cache-to', {
+            alias: 'o',
+            type: 'string',
+            describe: 'Append all translated snippets to the given tablet file',
+            requiresArg: true,
+            default: undefined,
+          }),
       wrapHandler(async (args) => {
         const absAssemblies = (args.ASSEMBLY.length > 0 ? args.ASSEMBLY : ['.']).map((x) => path.resolve(x));
-        const absOutput = path.resolve(args.output);
-        const result = await infuse(absAssemblies, args.TABLET, {
-          outputFile: absOutput,
-          log: args.log,
-          tabletOutputFile: args.TABLET,
+        const cacheToFile = fmap(args['cache-to'], path.resolve);
+        const result = await infuse(absAssemblies, {
+          logFile: args['log-file'],
+          cacheToFile: cacheToFile,
         });
 
         let totalTypes = 0;
@@ -131,10 +126,10 @@ function main() {
             describe: 'Assembly or directory to extract from',
           })
           .option('output', {
-            alias: 'o',
             type: 'string',
-            describe: 'Output file where to store the sample tablets',
-            default: DEFAULT_TABLET_NAME,
+            describe: 'Additional output file where to store translated samples (deprecated, alias for --cache-to)',
+            requiresArg: true,
+            default: undefined,
           })
           .option('compile', {
             alias: 'c',
@@ -173,9 +168,32 @@ function main() {
             alias: 'C',
             type: 'string',
             // eslint-disable-next-line prettier/prettier
-            describe: 'Reuse translations from the given tablet file if the snippet and type definitions did not change',
+            describe:
+              'Reuse translations from the given tablet file if the snippet and type definitions did not change',
             requiresArg: true,
             default: undefined,
+          })
+          .option('cache-to', {
+            alias: 'o',
+            type: 'string',
+            describe: 'Append all translated snippets to the given tablet file',
+            requiresArg: true,
+            default: undefined,
+          })
+          .conflicts('cache-to', 'output')
+          .option('cache', {
+            alias: 'k',
+            type: 'string',
+            describe: 'Alias for --cache-from and --cache-to together',
+            requiresArg: true,
+            default: undefined,
+          })
+          .conflicts('cache', 'cache-from')
+          .conflicts('cache', 'cache-to')
+          .option('trim-cache', {
+            alias: 'T',
+            type: 'boolean',
+            describe: 'Remove translations that are not referenced by any of the assemblies anymore from the cache',
           })
           .option('strict', {
             alias: 'S',
@@ -197,30 +215,26 @@ function main() {
         // compilerhost. Have to make all file references absolute before we chdir
         // though.
         const absAssemblies = (args.ASSEMBLY.length > 0 ? args.ASSEMBLY : ['.']).map((x) => path.resolve(x));
-        const absOutput = path.resolve(args.output);
-        const absCache = fmap(args['cache-from'], path.resolve);
+
+        const absCacheFrom = fmap(args.cache ?? args['cache-from'], path.resolve);
+        const absCacheTo = fmap(args.cache ?? args['cache-to'] ?? args.output, path.resolve);
+
         if (args.directory) {
           process.chdir(args.directory);
         }
 
-        let result: ExtractResult;
-        if (args.infuse) {
-          result = await extractAndInfuse(absAssemblies, {
-            outputFile: absOutput,
-            includeCompilerDiagnostics: !!args.compile,
-            validateAssemblies: args['validate-assemblies'],
-            only: args.include,
-            cacheTabletFile: absCache,
-          });
-        } else {
-          result = await extractSnippets(absAssemblies, {
-            outputFile: absOutput,
-            includeCompilerDiagnostics: !!args.compile,
-            validateAssemblies: args['validate-assemblies'],
-            only: args.include,
-            cacheTabletFile: absCache,
-          });
-        }
+        const extractOptions: ExtractOptions = {
+          includeCompilerDiagnostics: !!args.compile,
+          validateAssemblies: args['validate-assemblies'],
+          only: args.include,
+          cacheFromFile: absCacheFrom,
+          cacheToFile: absCacheTo,
+          trimCache: args['trim-cache'],
+        };
+
+        const result = args.infuse
+          ? await extractAndInfuse(absAssemblies, extractOptions)
+          : await extractSnippets(absAssemblies, extractOptions);
 
         handleDiagnostics(result.diagnostics, args.fail, result.tablet.count);
       }),
@@ -280,6 +294,30 @@ function main() {
               })
             : Object.values(TargetLanguage);
         return transliterateAssembly(assemblies, languages, args);
+      }),
+    )
+    .command(
+      'trim-cache <TABLET> [ASSEMBLY..]',
+      'Retain only those snippets in the cache which occur in one of the given assemblies',
+      (command) =>
+        command
+          .positional('TABLET', {
+            type: 'string',
+            required: true,
+            describe: 'Language tablet to trim',
+          })
+          .positional('ASSEMBLY', {
+            type: 'string',
+            string: true,
+            default: new Array<string>(),
+            describe: 'Assembly or directory to search',
+          })
+          .demandOption('TABLET'),
+      wrapHandler(async (args) => {
+        await trimCache({
+          cacheFile: args.TABLET,
+          assemblyLocations: args.ASSEMBLY,
+        });
       }),
     )
     .command(